Markdown fixes: table formatting (#601)
* fix: do not add new lines in markdown cells

* fix: markdown tables can only have one header

* fix: add a space after text before a paragraph in a cell

Paragraph is a block level element so we would normally have a new line before
it. However, here we are in a markdown cell and can't have new lines, so add a
space. Otherwise words will get concatenated.

* fix: match maximum cell count on each row

Markdown does not support colspan, but at least this way we don't lose any cell
data.

* fix: cells always need to append vertical bars

Currently there was a scenario where if a cell only contains a single <p> with
some text but no text directly, then vertical bars would not get appended for
that cells.

* Fix table processing tests and add a few more

Added the following table tests in text format:
- removing new lines in cells,
- only allowing a single header row,
- handling colspan by appending columns.

* fix: remove row span attribute once it is no longer useful